Added Iframe Support
now the dialogs can take content as a string or dom element. or instead
it can take a url. when a url is specified the main content div is
replaced by an iframe and the source of the iframe is set to the url.

along with iframe support the onload event has can now be used in the
jsDialog, but only if url is specified.